Sorting fix, --no-start options

+++ lib/spark/sort.rb
@@ -1,68 +1,92 @@
 # require "algorithms"
+require "tmpdir"
 
 module Spark
-  module ExternalSorter
+  class ExternalSorter  
 
     include Spark::Helper::System
 
-    # Items from GC can not be completely free so we need some reserve
-    MEMORY_RESERVE = 20
+    # Items from GC cannot be destroyed so #make_parts need some reserve
+    MEMORY_RESERVE = 50 # %
 
-    # How many items will be evaluate from iterator
-    SLICE_SIZE = 10
+    # How big will be chunk for adding new memory because GC not cleaning
+    # immediately un-referenced variables
+    MEMORY_FREE_CHUNK = 10 # %
 
+    # How many items will be evaluate from iterator at start
+    START_SLICE_SIZE = 10
+
+    # Maximum of slicing. Memory control can be avoided by large value.
+    MAX_SLICE_SIZE = 10_000
+
+    # How many values will be taken from each enumerator.
     EVAL_N_VALUES = 10
 
-    attr_reader :total_memory, :memory_limit, :serializer
+    # Default key function
+    KEY_FUNCTION = lambda{|item| item}
+
+    attr_reader :total_memory, :memory_limit, :memory_chunk, :serializer
 
     def initialize(total_memory, serializer)
       @total_memory = total_memory
-      @memory_limit = total_memory * (100-MEMORY_RESERVE)
+      @memory_limit = total_memory * (100-MEMORY_RESERVE)    / 100
+      @memory_chunk = total_memory * (100-MEMORY_FREE_CHUNK) / 100
       @serializer   = serializer
     end
 
-    def sort_by(iterator, key_function)
-      return to_enum(__callee__, iterator) unless block_given?
+    def add_memory!
+      @memory_limit += memory_chunk
+    end
 
-      init_temp_folder
+    def sort_by(iterator, key_function=KEY_FUNCTION)
+      return to_enum(__callee__, iterator, key_function) unless block_given?
 
+      create_temp_folder
+
+      # Make N sorted enumerators
       parts = make_parts(iterator, key_function)
 
       return [] if parts.empty?
 
-      eval_n_values = EVAL_N_VALUES
-      eval_n_values = parts.size if parts.size > EVAL_N_VALUES
+      # Need new key function because items have new structure
+      # From: [1,2,3] to [[1, Enumerator],[2, Enumerator],[3, Enumerator]]
+      key_function_with_enum = lambda{|(key, _)| key_function[key]}
 
       heap  = []
       enums = []
 
+      # Load first items to heap
       parts.each do |part|
-        eval_n_values.times {
+        EVAL_N_VALUES.times {
           begin
             heap << [part.next, part]
           rescue StopIteration
+            break
           end
         }
       end
 
+      # Parts can be empty but heap not
       while parts.any? || heap.any?
-          heap.sort_by!(&key_function)
+        heap.sort_by!(&key_function_with_enum)
 
-          eval_n_values.times {
-            break if heap.empty?
+        # Since parts are sorted and heap contains EVAL_N_VALUES method
+        # can add EVAL_N_VALUES items to the result
+        EVAL_N_VALUES.times {
+          break if heap.empty?
 
-            item, enum = heap.shift
+          item, enum = heap.shift
+          enums << enum
 
-            yield item
+          yield item
+        }
 
-            enums << enum
-          }
-
+        # Add new element to heap from part of which was result item
         while (enum = enums.shift)
           begin
             heap << [enum.next, enum]
           rescue StopIteration
-            data.delete(enum)
+            parts.delete(enum)
             enums.delete(enum)
           end
         end
@@ -74,34 +98,55 @@ module Spark
 
     private
 
-      def init_temp_folder
+      def create_temp_folder
         @dir = Dir.mktmpdir
       end
 
       def destroy_temp_folder
-        FileUtils.remove_entry_secure(@dir)
+        FileUtils.remove_entry_secure(@dir) if @dir
       end
 
+      # New part is created when current part exceeds memory limit (is variable)
+      # Every new part have more memory because of ruby GC
       def make_parts(iterator, key_function)
+        slice = START_SLICE_SIZE
+
         parts = []
-        tmp   = []
+        part  = []
 
-        iterator.each_slice(SLICE_SIZE).with_index do |part, i|
-          tmp += part
+        loop do
+          begin
+            # Enumerator does not have slice method
+            slice.times { part << iterator.next }
+          rescue StopIteration
+            break
+          end
 
+          # Carefully memory_limit is variable 
           if memory_usage > memory_limit
-            tmp.sort_by!(&block)
-            file = Tempfile.new("part_#{i}", @dir)
-            serializer.dump(tmp, file)
-            parts << serializer.load_as_enum(file)
+            # Sort current part with origin key_function
+            part.sort_by!(&key_function)
+            # Tempfile for current part
+            # will be destroyed on #destroy_temp_folder
+            file = Tempfile.new("part", @dir)
+            serializer.dump(part, file)
+            # Peek is at the end of file
+            file.seek(0)
+            parts << serializer.load(file)
 
-            # Some memory will be released
-            tmp.clear
+            # Some memory will be released but not immediately
+            # need some new memory for start
+            part.clear
+            add_memory!
+          else
+            slice = [slice*2, MAX_SLICE_SIZE].min
           end
         end
 
-        if tmp.any?
-          parts << tmp.each
+        # Last part which is not in the file
+        if part.any?
+          part.sort_by!(&key_function)
+          parts << part.each
         end
 
         parts
